+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * export initial tempo and meter
 * predict General Midi program numbers for all orchestral strings
 * added command line argument for MuseScore (to work around MIDI import of piano tracks)
+* ignore note_on events at the very end of a region
 
 ## 0.2
 * fixed import of tracks that start later in piece rather than at very beginning