+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Change log
 
-## next
+## 0.3
 * export initial tempo and meter
 * predict General Midi program numbers for all orchestral strings
 * added command line argument for MuseScore (to work around MIDI import of piano tracks)
+* added command line argument to omit track name text that is inside parentheses () or square brackets []
 * ignore note_on events at the very end of a region
 * send note_off events for any notes on at the end of a region